Added an SPF handler + moved handlers to separate file
(Pushed into this branch so I don't need to rewrite logging again for those new statements)
