+++ main.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	"bytes"
-	"errors"
 	"os"
 	"os/signal"
 	"syscall"
@@ -10,54 +8,7 @@ import (
 	"github.com/gopistolet/gopistolet/helpers"
 	"github.com/gopistolet/gopistolet/log"
 	"github.com/gopistolet/gopistolet/mta"
-	"github.com/sloonz/go-maildir"
 )
-
-var mailDir *maildir.Maildir
-
-func handleMailDir(state *mta.State) {
-	err := errors.New("")
-
-	// Open maildir if it's not yet open
-	if mailDir == nil {
-
-		// Open a maildir. If it does not exist, create it.
-		mailDir, err = maildir.New("./maildir", true)
-		if err != nil {
-			log.Errorln(err)
-		}
-	}
-
-	dataReader := bytes.NewReader(state.Data)
-
-	// Save mail in maildir
-	filename, err := mailDir.CreateMail(dataReader)
-	if err != nil {
-		//log.Println(err)
-		log.WithFields(log.Fields{
-			"SessionId": state.SessionId.String(),
-		}).Error(err)
-	} else {
-		//log.Println("Maildir: mail written to file: " + filename)
-		log.WithFields(log.Fields{
-			"SessionId": state.SessionId.String(),
-		}).Info("Maildir: mail written to file: " + filename)
-	}
-}
-
-func mail(state *mta.State) {
-	log.Debugf("From: %s\n", state.From.Address)
-	log.Debugf("To: ")
-	for i, to := range state.To {
-		log.Printf("%s", to.Address)
-		if i != len(state.To)-1 {
-			log.Printf(",")
-		}
-	}
-	log.Debugf("CONTENT_START:\n")
-	log.Debugf("%s\n", string(state.Data))
-	log.Debugf("CONTENT_END\n")
-}
 
 type Chain struct {
 	handlers []mta.Handler
@@ -95,6 +46,7 @@ func main() {
 	mta := mta.NewDefault(c,
 		&Chain{handlers: []mta.Handler{
 			mta.HandlerFunc(mail),
+			mta.HandlerFunc(handleSPF),
 			mta.HandlerFunc(handleMailDir),
 		}})
 	go func() {
